+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - [Imp] Configuration display in admin: better handling of secret value obfuscation (#93)
 - [Imp] No "Add filament" buttons if there is no materials, brands and shops (#90)
 - [Imp] Allow changelog page to be accessible for non-logged users (#94)
+- [Imp] Hompage's count per material chart no use root material for grouping (#96)
 - [Fix] Filament statistics page errors when no filaments (#91)
 - [Fix] No link to filament statistics when no filaments (#95)
 - [Tec] Add `versionSuffix` config setting (#92)